feat: add env schema to base project template
