--- template/addons/prisma/next.config.js
@@ -1,8 +1,0 @@
-const { env } = require("./src/server/env");
-
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  reactStrictMode: true,
-};
-
-module.exports = nextConfig;
