+++ template/base/next.config.js
@@ -1,6 +1,8 @@
+const { env } = require("./src/server/env");
+
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-}
+};
 
-module.exports = nextConfig
+module.exports = nextConfig;
