+++ template/base/package.json
@@ -11,7 +11,8 @@
   "dependencies": {
     "next": "12.2.0",
     "react": "18.2.0",
-    "react-dom": "18.2.0"
+    "react-dom": "18.2.0",
+    "zod": "^3.17.3"
   },
   "devDependencies": {
     "@types/node": "18.0.0",
