+++ template/base/src/server/env.js
@@ -1,0 +1,29 @@
+// @ts-check
+/**
+ * This file is included in `/next.config.js` which ensures the app isn't built with invalid env vars.
+ * It has to be a `.js`-file to be imported there.
+ */
+const { envSchema } = require("./env-schema");
+
+const env = envSchema.safeParse(process.env);
+
+const formatErrors = (
+  /** @type {import('zod').ZodFormattedError<Map<string,string>,string>} */
+  errors,
+) =>
+  Object.entries(errors)
+    .map(([name, value]) => {
+      if (value && "_errors" in value)
+        return `${name}: ${value._errors.join(", ")}\n`;
+    })
+    .filter(Boolean);
+
+if (!env.success) {
+  console.error(
+    "❌ Invalid environment variables:\n",
+    ...formatErrors(env.error.format()),
+  );
+  process.exit(1);
+}
+
+module.exports.env = env.data;
